+++ packages/vtable-plugins/demo/menu.ts
@@ -12,6 +12,10 @@ export const menus = [
     name: 'excel-import'
   },
   {
+    path: 'excel-import',
+    name: 'issue-5227-leading-blank-row'
+  },
+  {
     path: 'filter',
     name: 'filter'
   },
@@ -22,6 +26,10 @@ export const menus = [
   {
     path: 'filter',
     name: 'issue-5137'
+  },
+  {
+    path: 'filter',
+    name: 'issue-4625-filter-popup-position'
   },
   {
     path: 'filter',
@@ -84,6 +92,10 @@ export const menus = [
     name: 'rotate-table'
   },
   {
+    path: 'rotate-table',
+    name: 'issue-5235-rotate-wheel'
+  },
+  {
     path: 'table-series-number',
     name: 'table-series-number'
   },
@@ -128,6 +140,10 @@ export const menus = [
     name: 'master-detail-plugin9'
   },
   {
+    path: 'master-detail-plugin',
+    name: 'issue-5185-set-records-expand'
+  },
+  {
     menu: 'pivot-plugin',
     children: [
       {
@@ -139,6 +155,14 @@ export const menus = [
   {
     path: 'context-menu',
     name: 'context-menu'
+  },
+  {
+    path: 'context-menu',
+    name: 'issue-5215-context-menu-canvas'
+  },
+  {
+    path: 'context-menu',
+    name: 'issue-5214-reverse-selected-row-delete'
   },
   {
     path: 'table-export',